feature: Added 'duplicate_skip' option to functions in controller.py
Added a 'duplicate_skip' parameter to several functions in the controller.py file. This allows the user to specify whether DuplicateKeyErrors should be skipped or not. It aims to enhance the functionality of the Retry decorator by providing an option to ignore DuplicateKeyErrors, thus improving application efficiency and error handling.